+++ packages/web/components/pool-detail/index.ts
@@ -1,0 +1,2 @@
+export * from "./concentrated-liquidity-pool";
+export * from "./share-pool";
